Enhanced the 'visualize_query' tool to support specifying visualization types and shapes, improving flexibility.
