+++ src/platform/packages/shared/kbn-visualization-utils/index.ts
@@ -11,7 +11,7 @@ export { getTimeZone } from './src/get_timezone';
 export { getLensAttributesFromSuggestion } from './src/get_lens_attributes';
 export { TooltipWrapper } from './src/tooltip_wrapper';
 export { useDebouncedValue } from './src/debounced_value';
-export { ChartType } from './src/types';
+export { ChartType, LensVisualizationType } from './src/types';
 export { getDatasourceId } from './src/get_datasource_id';
 export { mapVisToChartType } from './src/map_vis_to_chart_type';
 export { computeInterval } from './src/compute_interval';
