+++ x-pack/platform/plugins/shared/lens/public/types.ts
@@ -50,7 +50,7 @@ import type { DraggingIdentifier, DragDropIdentifier, DropType } from '@kbn/dom-
 import type { AccessorConfig } from '@kbn/visualization-ui-components';
 import type { ChartSizeEvent } from '@kbn/chart-expressions-common';
 import type { AlertRuleFromVisUIActionData } from '@kbn/alerts-ui-shared';
-import type { DateRange, LayerType, SortingHint } from '../common/types';
+import type { DateRange, LayerType, SortingHint, VisualizationType } from '../common/types';
 import type {
   LensSortActionData,
   LensResizeActionData,
@@ -976,45 +976,6 @@ export interface FramePublicAPI {
   activeData?: Record<string, Datatable>;
   dataViews: DataViewsState;
   forceDSL?: boolean;
-}
-
-/**
- * A visualization type advertised to the user in the chart switcher
- */
-export interface VisualizationType {
-  /**
-   * Unique id of the visualization type within the visualization defining it
-   */
-  id: string;
-  /**
-   * Icon used in the chart switcher
-   */
-  icon: IconType;
-  /**
-   * Visible label used in the chart switcher and above the workspace panel in collapsed state
-   */
-  label: string;
-  description: string;
-  /**
-   * Optional label used in visualization type search if chart switcher is expanded and for tooltips
-   */
-  fullLabel?: string;
-  /**
-   * Priority of the visualization for sorting in chart switch
-   * Lower number means higher priority (aka top of list).
-   *
-   */
-  sortPriority: number;
-  /**
-   * Indicates if visualization is in the experimental stage.
-   */
-  showExperimentalBadge?: boolean;
-  /**
-   * Indicates if visualization is deprecated.
-   */
-  isDeprecated?: boolean;
-  subtypes?: string[];
-  getCompatibleSubtype?: (seriesType?: string) => string | undefined;
 }
 
 export interface VisualizationDisplayOptions {
