+++ x-pack/platform/plugins/shared/lens/public/visualizations/tagcloud/constants.ts
@@ -5,13 +5,8 @@
  * 2.0.
  */
 
-import { i18n } from '@kbn/i18n';
 import type { $Values } from '@kbn/utility-types';
 import { Orientation } from '@kbn/expression-tagcloud-plugin/common';
-
-export const TAGCLOUD_LABEL = i18n.translate('xpack.lens.tagcloud.label', {
-  defaultMessage: 'Tag cloud',
-});
 
 export const DEFAULT_STATE = {
   maxFontSize: 72,
