+++ x-pack/platform/plugins/shared/lens/public/visualizations/xy/xy_suggestions.test.ts
@@ -8,7 +8,7 @@
 import { getSuggestions } from './xy_suggestions';
 import type { TableSuggestionColumn, VisualizationSuggestion, TableSuggestion } from '../../types';
 import type { State, XYState, XYAnnotationLayerConfig, XYDataLayerConfig } from './types';
-import { visualizationSubtypes } from './types';
+import { visualizationSubtypes } from '../../../common/visualizations/xy_visualization';
 import { generateId } from '../../id_generator';
 import { type PaletteOutput, DEFAULT_COLOR_MAPPING_CONFIG } from '@kbn/coloring';
 import { LayerTypes } from '@kbn/expression-xy-plugin/public';
