+++ x-pack/platform/plugins/shared/observability_ai_assistant/common/functions/visualize_esql.ts
@@ -9,6 +9,10 @@ export enum VisualizeESQLUserIntention {
   generateQueryOnly = 'generateQueryOnly',
   executeAndReturnResults = 'executeAndReturnResults',
   visualizeAuto = 'visualizeAuto',
+  visualize = 'visualize',
+}
+
+export enum LegacyVisualizeESQLUserIntention {
   visualizeXy = 'visualizeXy',
   visualizeBar = 'visualizeBar',
   visualizeLine = 'visualizeLine',
