+++ x-pack/platform/plugins/shared/observability_ai_assistant/public/index.ts
@@ -60,6 +60,7 @@ export {
 export {
   VisualizeESQLUserIntention,
   VISUALIZE_ESQL_USER_INTENTIONS,
+  LegacyVisualizeESQLUserIntention,
 } from '../common/functions/visualize_esql';
 
 export {
