+++ x-pack/solutions/observability/plugins/observability_ai_assistant_app/common/functions/visualize_esql.ts
@@ -5,33 +5,9 @@
  * 2.0.
  */
 import type { FromSchema } from 'json-schema-to-ts';
-import {
-  VISUALIZE_ESQL_USER_INTENTIONS,
-  VISUALIZE_QUERY_FUNCTION_NAME,
-} from '@kbn/observability-ai-assistant-plugin/common';
 import type { ESQLRow } from '@kbn/es-types';
 import type { DatatableColumn } from '@kbn/expressions-plugin/common';
-
-export const visualizeESQLFunction = {
-  name: VISUALIZE_QUERY_FUNCTION_NAME,
-  isInternal: true,
-  description: 'Use this function to visualize charts for ES|QL queries.',
-  descriptionForUser: 'Use this function to visualize charts for ES|QL queries.',
-  parameters: {
-    type: 'object',
-    properties: {
-      query: {
-        type: 'string',
-      },
-      intention: {
-        type: 'string',
-        enum: VISUALIZE_ESQL_USER_INTENTIONS,
-      },
-    },
-    required: ['query', 'intention'],
-  } as const,
-  contexts: ['core'],
-};
+import type { visualizeESQLFunction } from '../../server/functions/visualize_esql';
 
 export interface VisualizeQueryResponsev0 {
   content: DatatableColumn[];
